pyproject.toml
[tool.poetry]
name = "rag-textbook"
version = "0.1.0"
description = "rag for textbooks"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = 'src'}
]

[tool.poetry.dependencies]
python = "^3.12"
textract = "^1.6.5"
langchain-community = "^0.3.1"
pypdf = "^5.0.1"
argcomplete = "~=1.10.0"
beautifulsoup4 = "~=4.8.0"
chardet = ">=3,<4"
docx2txt = "~=0.8"
extract-msg = "<=0.29.*"  # Last version with Python 2 support
python-pptx = "~=0.6.18"
SpeechRecognition = "~=3.8.1"
xlrd = "~=1.2.0"
pdfminer-six = "20191110"
setuptools = "^75.1.0"
asyncio = "^3.4.3"
llamaapi = "^0.1.36"
huggingface-hub = "^0.26.1"
transformers = "^4.45.2"
torch = "^2.5.0"
cohere = "^5.11.1"
requests = "^2.32.3"
openai = "^1.52.2"
spacy = "3.7.3"
tiktoken = "^0.8.0"
bertopic = "^0.16.4"
langchain-experimental = "^0.3.2"
langchain-openai = "^0.2.3"
sentence-transformers = "^3.2.1"
pyyaml = "^6.0.2"
pinecone = "^5.3.1"
six = ">=1.12.0,<1.13.0"
pinecone-client = "^5.0.1"
streamlit = "^1.40.1"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
